Move services to the root of src/restate-app
