+++ typescript/food-ordering/app/src/restate-app/external/driver_mobile_app_sim_utils.ts
